Initial commit. Delete options initially for Warp 10 Accelerator, extended for all deletes

+++ warp10/src/main/java/io/warp10/standalone/StandaloneDeleteHandler.java
@@ -306,7 +306,7 @@ public class StandaloneDeleteHandler extends AbstractHandler {
       response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Standalone version does not support the '" + Constants.HTTP_PARAM_MINAGE + "' parameter in delete requests.");
       return;
     }
-    
+        
     boolean dryrun = null != request.getParameter(Constants.HTTP_PARAM_DRYRUN);
     
     File loggingFile = null;
@@ -457,6 +457,18 @@ public class StandaloneDeleteHandler extends AbstractHandler {
       if (Long.MIN_VALUE != start || Long.MAX_VALUE != end) {
         hasRange = true;
       }
+
+      boolean nodata = null != request.getParameter(Constants.HTTP_PARAM_NODATA);
+      
+      if (nodata && !Constants.DELETE_NODATA_SUPPORT) {
+        response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Parameter " + Constants.HTTP_PARAM_NODATA + " cannot be used as nodata support is not enabled.");
+        return;        
+      }
+      
+      if (nodata && hasRange) {
+        response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Parameter " + Constants.HTTP_PARAM_NODATA + " can only be set if no range is specified.");
+        return;
+      }
       
       if (start > end) {
         response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Invalid time range specification.");
@@ -506,6 +518,25 @@ public class StandaloneDeleteHandler extends AbstractHandler {
       lblsSels.add(labelsSelectors);
       
       DirectoryRequest drequest = new DirectoryRequest();
+      
+      Long activeAfter = null == request.getParameter(Constants.HTTP_PARAM_ACTIVEAFTER) ? null : Long.parseLong(request.getParameter(Constants.HTTP_PARAM_ACTIVEAFTER));
+      Long quietAfter = null == request.getParameter(Constants.HTTP_PARAM_QUIETAFTER) ? null : Long.parseLong(request.getParameter(Constants.HTTP_PARAM_QUIETAFTER));
+
+      if (!Constants.DELETE_ACTIVITY_SUPPORT) {
+        if (null != activeAfter || null != quietAfter) {
+          response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Activity based selection is disabled by configuration.");
+          return;
+        }
+      }
+      
+      if (null != activeAfter) {
+        drequest.setActiveAfter(activeAfter);
+      }
+      
+      if (null != quietAfter) {
+        drequest.setQuietAfter(quietAfter);
+      }
+      
       drequest.setClassSelectors(clsSels);
       drequest.setLabelsSelectors(lblsSels);
 
@@ -536,7 +567,9 @@ public class StandaloneDeleteHandler extends AbstractHandler {
               continue;
             }
           }
-          localCount = this.storeClient.delete(writeToken, metadata, start, end);
+          if (!nodata) {
+            localCount = this.storeClient.delete(writeToken, metadata, start, end);
+          }
         }
 
         //
